instead of get collegeid incase of student/room insertion from client , get it from database using id which is stored in auth context

+++ src/main/java/com/example/SmartSeatBackend/entity/Students.java
@@ -51,6 +51,6 @@ public class Students {
     private String imgUrl;
 
     @Column(name = "college_id")
-    private Integer collegeId;
+    private Long collegeId;
 
 }
